PathManager 추가 PathManager는 프로그램의 실행 상대경로를 구함

+++ WinApi2D/CMissile.cpp
@@ -40,6 +40,6 @@ void CMissile::SetDir(fVec2 vec)
 
 void CMissile::SetDir(float theta)
 {
-	m_fvDir.x = cos(theta);
-	m_fvDir.y = sin(theta);
+	m_fvDir.x = (float)cos(theta);
+	m_fvDir.y = (float)sin(theta);
 }
